T25 review round 1: close boundary replay gaps
Close the remaining PFC lane, controller, hysteresis, ACK sizing, and
capability boundaries found by the post-fix audit. Count blocked rate
tokens exactly once and make DRR reachability and trace bounds explicit.

Reject unrepresentable byte occupancy for every queue policy, preserve
RED transition order, and replay the new boundary rows in Lean. Raise the
WRR replay fuel to cover a full wrapped scan.

Tests:
- focused executor and lowering regressions: 72 passed
- LeanGuard mechanism campaign: 11/11
- LeanGuard AQM campaign: 8/8
- lake build: 58 jobs
- all four Clippy configurations
- cargo xtask audit
- frozen pre-optimization hashes: 2/2

+++ executor/src/mechanism_trace.rs
@@ -95,7 +95,7 @@ pub struct DrrTransitionRecord {
     pub quanta_bytes: Vec<u64>,
     pub before_deficits_bytes: Vec<u64>,
     pub before_current_class: u64,
-    pub scan_steps: u64,
+    pub scan_steps: u128,
     pub eligible_packets: Vec<SchedulerPacket>,
     pub selected_payload: PayloadId,
     pub after_deficits_bytes: Vec<u64>,
